requirements.txt
flask
flask_sqlalchemy
flask_login
python-dotenv
gunicorn
